Restructure network tests directory for clarity
Test directory contains the programs that need to run on the Pi and on
the Remote host
The directory is broken down into which test you want to run (packet
size or frequency), then has subdirectories containing the Pi code and
remote host code
